découpe de User + changement librairie PDF
