+++ .flutter-plugins
@@ -5,7 +5,6 @@ connectivity_for_web=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\con
 connectivity_macos=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\connectivity_macos-0.1.0+7\\
 device_info=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\device_info-1.0.0\\
 flutter_email_sender=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\flutter_email_sender-4.0.0\\
-flutter_full_pdf_viewer=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\flutter_full_pdf_viewer-1.0.6\\
 flutter_secure_storage=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\flutter_secure_storage-3.3.5\\
 flutter_siri_suggestions=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\flutter_siri_suggestions-1.0.2+hotfix1\\
 flutter_statusbarcolor=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\flutter_statusbarcolor-0.2.3\\
@@ -25,6 +24,7 @@ shared_preferences_linux=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\
 shared_preferences_macos=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\shared_preferences_macos-0.0.1+11\\
 shared_preferences_web=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\shared_preferences_web-0.1.2+7\\
 shared_preferences_windows=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\shared_preferences_windows-0.0.2+2\\
+syncfusion_flutter_pdfviewer=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\syncfusion_flutter_pdfviewer-18.4.39-beta\\
 url_launcher=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\url_launcher-5.7.10\\
 url_launcher_linux=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\url_launcher_linux-0.0.1+4\\
 url_launcher_macos=C:\\src\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\url_launcher_macos-0.0.1+9\\
