+++ .packages
@@ -1,4 +1,4 @@
-# Generated by pub on 2021-01-28 17:27:24.940048.
+# Generated by pub on 2021-02-01 15:26:06.460965.
 _fe_analyzer_shared:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/_fe_analyzer_shared-7.0.0/lib/
 about:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/about-1.4.0/lib/
 analyzer:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/analyzer-0.39.17/lib/
@@ -30,7 +30,6 @@ ffi:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/ffi-0.1.3/lib/
 file:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/file-5.2.1/lib/
 flutter:file:///C:/src/flutter/packages/flutter/lib/
 flutter_email_sender:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_email_sender-4.0.0/lib/
-flutter_full_pdf_viewer:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_full_pdf_viewer-1.0.6/lib/
 flutter_localizations:file:///C:/src/flutter/packages/flutter_localizations/lib/
 flutter_markdown:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_markdown-0.5.2/lib/
 flutter_phoenix:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_phoenix-0.1.0/lib/
@@ -97,9 +96,11 @@ stack_trace:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/stack_trac
 stream_channel:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.1.0-nullsafety.1/lib/
 string_scanner:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/string_scanner-1.1.0-nullsafety.1/lib/
 syncfusion_flutter_calendar:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/syncfusion_flutter_calendar-18.4.35/lib/
-syncfusion_flutter_core:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/syncfusion_flutter_core-18.4.35+1/lib/
+syncfusion_flutter_core:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/syncfusion_flutter_core-18.4.39/lib/
 syncfusion_flutter_datagrid:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/syncfusion_flutter_datagrid-18.4.35-beta/lib/
 syncfusion_flutter_datepicker:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/syncfusion_flutter_datepicker-18.4.35-beta/lib/
+syncfusion_flutter_pdf:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/syncfusion_flutter_pdf-18.4.39-beta/lib/
+syncfusion_flutter_pdfviewer:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/syncfusion_flutter_pdfviewer-18.4.39-beta/lib/
 syncfusion_localizations:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/syncfusion_localizations-18.4.35/lib/
 synchronized:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/synchronized-2.2.0+2/lib/
 term_glyph:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/term_glyph-1.2.0-nullsafety.1/lib/
